Moved camera-preview into separate class
