+++ app/src/androidTest/java/com/github/bgabriel998/softwaredevproject/CameraPreviewTest.java
@@ -41,7 +41,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 
 
-
 @RunWith(AndroidJUnit4.class)
 public class CameraPreviewTest implements LifecycleOwner, ImageReader.OnImageAvailableListener, Consumer<SurfaceRequest.Result> {
     @Rule
@@ -166,5 +165,13 @@ public class CameraPreviewTest implements LifecycleOwner, ImageReader.OnImageAva
     public Lifecycle getLifecycle() {
         return registry;
     }
+
+//    @Test
+//    public void aspectRatioTest(){
+//        Context context = ApplicationProvider.getApplicationContext();
+//        CameraPreview preview = new CameraPreview(context);
+//        assertEquals(preview.aspectRatio(4, 3), AspectRatio.RATIO_4_3);
+//        assertEquals(preview.aspectRatio(16, 9), AspectRatio.RATIO_16_9);
+//    }
 
 }
